src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.35.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1